refactor(auth): 共享值类型迁入 types.py（删旧契约的前置）
